# server/openapi.yaml
openapi: 3.1.0
info:
  title: 采购看板 API
  version: 0.1.0
  description: |
    Procurement Kanban backend API.
    Use `/auth/login` (human) or hold a long-lived token issued by `/auth/agent-token` (machine / MCP).
    All authenticated requests must include `Authorization: Bearer <token>`.
servers:
  - url: http://localhost:3001/api
    description: Local dev
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [auth]
      summary: Login with email + password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Logged in
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '401': { description: Invalid credentials }
...
  /auth/agent-token:
    post:
      tags: [auth]
      summary: Issue a long-lived agent JWT (for MCP / programmatic agents)
      description: |
        Caller must be authenticated as a human user (i.e. logged in via /auth/login).
        Returned token carries `{ agent: true, scope, agentName }` claims and a TTL up to 365 days.
        Currently only `scope=rw` is enforced; `ro` reserved for future RO middleware.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, description: 'Friendly token label', default: agent }
                scope: { type: string, enum: [ro, rw], default: rw }
                days: { type: integer, minimum: 1, maximum: 365, default: 365 }
      responses:
        '200':
          description: Token issued
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  expiresInDays: { type: integer }
                  scope: { type: string }
                  agentName: { type: string }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        email: { type: string }
        role: { type: string }
        departmentId: { type: string, nullable: true }
...